+++ CHANGELOG.md
@@ -13,6 +13,9 @@ You can also check the [release page](https://github.com/visualize-admin/visuali
   - Cascading filters are not stuck anymore in the loading mode in some cases
 - Maintenance
   - GQL debug panel now includes queries fired through SPARQLClientStream (e.g. hierarchies) and CONSTRUCT queries
+  - Configurator and interactive filters debug panel is now displayed if `flag__debug` is set to true
+- Docs
+  - Added chart preview via API section to the documentation
 
 # [3.22.8] - 2023-09-29